SIACASA/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
coverage==7.8.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
eventlet==0.39.1
faiss-cpu==1.11.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fonttools==4.58.2
frozenlist==1.6.2
fsspec==2025.5.1
greenlet==3.2.1
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.70.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.2
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-engineio==4.12.0
python-jose==3.4.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.1.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
alembic==1.12.1
aiofiles==23.2.1
